requirements.txt
langchain==0.0.257
openai==0.27.8
streamlit==1.25.0
python-dotenv==1.0.0